Update shipsimulation.m
Added list of waypoints and a waypoint index that increments each time the boat has reached its destination. Currently the boat will reach its first waypoint and start to move towards the second.

However, it will not update to the next waypoint, once it has reached the second, and will continue to twirl around the second waypoint as it stands.